create/edit/delete listing created

+++ routes.php
@@ -1,12 +1,20 @@
 <?php
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
-$router->get('/listing/edit/{id}', 'ListingController@edit');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get('/listing/edit/{id}', 'ListingController@edit', ['auth']);
+$router->get('listings/search', 'ListingController@search');
 $router->get('/listing/{id}', 'ListingController@show');
 
-$router->post('/listings', 'ListingController@store');
+$router->post('/listings', 'ListingController@store', ['auth']);
 
-$router->put('/listing/{id}', 'ListingController@update');
+$router->put('/listing/{id}', 'ListingController@update', ['auth']);
 
-$router->delete('/listing/{id}', 'ListingController@destroy');
+$router->delete('/listing/{id}', 'ListingController@destroy', ['auth']);
+
+$router->get('auth/register', 'UserController@create', ['guest']);
+$router->get('auth/login', 'UserController@login', ['guest']);
+
+$router->post('auth/register', 'UserController@store', ['guest']);
+$router->post('auth/logout', 'UserController@logout', ['auth']);
+$router->post('auth/login', 'UserController@authenticate', ['guest']);
